Move drag_select helper to conftest and fix onboarding test for WebKit
The onboarding flow test had the same drag_to() issue as the availability
test — Viselect listens for mousedown/mousemove/mouseup, not HTML5 drag
events. Moved drag_select() to conftest.py so both test files share it,
and replaced the drag_to + wait_for_timeout call in set_availability().
