+++ tests/test_end_to_end.py
@@ -20,6 +20,8 @@ from django.urls import reverse
 from playwright.sync_api import expect, BrowserContext
 from playwright.sync_api import Page
 
+from tests.conftest import drag_select
+
 from accounts.factories import UserFactory, UserAvailabilityFactory
 from accounts.models import CustomUser
 from home.factories import (
@@ -52,29 +54,6 @@ class TeamFormationTestData:
 
 
 RESULTS_PATTERN = re.compile(r"Showing \d+ of \d+ opportunities")
-
-
-def drag_select(page: Page, from_locator, to_locator) -> None:
-    """Simulate a mouse drag selection between two elements.
-
-    Uses explicit mouse events (mousedown/mousemove/mouseup) instead of
-    Playwright's drag_to(), which uses HTML5 drag events that libraries
-    like Viselect don't listen for — causing failures in WebKit.
-    """
-    from_locator.scroll_into_view_if_needed()
-    to_locator.scroll_into_view_if_needed()
-    box_start = from_locator.bounding_box()
-    box_end = to_locator.bounding_box()
-    start_x = box_start["x"] + box_start["width"] / 2
-    start_y = box_start["y"]
-    end_x = box_end["x"] + box_end["width"] / 2
-    end_y = box_end["y"] + box_end["height"]
-    page.mouse.move(start_x, start_y)
-    page.mouse.down()
-    page.mouse.move(end_x, end_y)
-    # Wait for viselect to visually mark the last cell before releasing
-    expect(to_locator).to_have_class(re.compile(r"\bselecting\b"))
-    page.mouse.up()
 
 
 @pytest.fixture
